+++ lib/translate/inquiry.h
@@ -23,8 +23,8 @@
 #include "third_party/spdk_defs/nvme_defs.h"
 
 namespace translator {
-void translate(absl::Span<const uint32_t>);
-StatusCode RawToScsiCommand(absl::Span<const uint32_t>,
+void translate(absl::Span<const uint8_t>);
+StatusCode RawToScsiCommand(absl::Span<const uint8_t>,
                             scsi_defs::InquiryCommand &);
 
 nvme_defs::IdentifyControllerData NvmeIdentifyController();
